+++ src/lib/server/api/v2/project_api.js
@@ -135,7 +135,7 @@ export async function getProjectDatasets(fetch, projectId) {
 
 	/** @type {import('fractal-components/types/api').DatasetV2[]} */
 	const datasets = await response.json();
-	datasets.sort((a, b) => (a.name < b.name ? -1 : a.name > b.name ? 1 : 0));
+
 	return datasets;
 }