+++ tests/v2/datasets/dataset_crud.spec.js
@@ -40,19 +40,19 @@ test('Create, update and delete a dataset [v2]', async ({ page, project }) => {
 		await verifyDatasetsCount(page, initialDatasetsCount + 1);
 		if (initialDatasetsCount === 1) {
 			let defaultDatasetRow = datasetTable.getByRole('row').nth(1);
-			expect(await defaultDatasetRow.getByRole('cell').nth(0).innerText()).toEqual('default');
+			await expect(defaultDatasetRow.getByRole('cell').nth(0)).toHaveText('default');
 		}
 		datasetRow = datasetTable.getByRole('row').nth(initialDatasetsCount + 1);
-		expect(await datasetRow.getByRole('cell').nth(0).innerText()).toEqual('test-dataset');
+		await expect(datasetRow.getByRole('cell').nth(0)).toHaveText('test-dataset');
 	});
 
 	await test.step('Filter dataset', async () => {
 		const datasetTable = page.getByRole('table').nth(0).locator('tbody');
-		expect(await datasetTable.getByRole('row').count()).toEqual(1);
+		await expect(datasetTable.getByRole('row')).toHaveCount(1);
 		await page.getByPlaceholder('Search dataset').fill('foo');
-		expect(await datasetTable.getByRole('row').count()).toEqual(0);
+		await expect(datasetTable.getByRole('row')).toHaveCount(0);
 		await page.getByPlaceholder('Search dataset').fill('test-dataset');
-		expect(await datasetTable.getByRole('row').count()).toEqual(1);
+		await expect(datasetTable.getByRole('row')).toHaveCount(1);
 	});
 
 	await test.step('Open dataset page', async () => {
@@ -159,6 +159,77 @@ test('Create, update and delete a dataset [v2]', async ({ page, project }) => {
 
 	await test.step('Cleanup', async () => {
 		fs.rmSync(exportedDatasetFile);
+	});
+
+	await test.step('Star datasets', async () => {
+		const name1 = 'test-dataset-renamed';
+		const name2 = 'test-dataset-renamed2';
+
+		let datasetTable = page.getByRole('table').nth(0);
+		await verifyDatasetsCount(page, 2);
+
+		// Initial state:
+		// ds1 -> not starred
+		// ds2 -> not starred
+		let dataset1 = datasetTable.getByRole('row').nth(1).getByRole('cell').nth(0);
+		let dataset2 = datasetTable.getByRole('row').nth(2).getByRole('cell').nth(0);
+
+		await expect(dataset1).toHaveText(name1);
+		await expect(dataset2).toHaveText(name2);
+
+		let starButton1 = dataset1.getByRole('button', { name: 'star dataset' });
+		let starIcon1 = starButton1.locator('i');
+		let starButton2 = dataset2.getByRole('button', { name: 'star dataset' });
+		let starIcon2 = starButton2.locator('i');
+
+		await expect(starIcon1).toHaveClass(/bi-star(?!-fill)/);
+		await expect(starIcon2).toHaveClass(/bi-star(?!-fill)/);
+
+		// Star ds2 and reload page
+		await starButton2.click();
+
+		await expect(starIcon1).toHaveClass(/bi-star(?!-fill)/);
+		await expect(starIcon2).toHaveClass(/bi-star-fill/);
+
+		await page.reload();
+
+		// After reload:
+		// ds2 -> starred
+		// ds1 -> not starred
+		dataset1 = datasetTable.getByRole('row').nth(1).getByRole('cell').nth(0);
+		dataset2 = datasetTable.getByRole('row').nth(2).getByRole('cell').nth(0);
+
+		await expect(dataset1).toHaveText(name2);
+		await expect(dataset2).toHaveText(name1);
+
+		starIcon1 = dataset1.getByRole('button', { name: 'star dataset' }).locator('i');
+		starIcon2 = dataset2.getByRole('button', { name: 'star dataset' }).locator('i');
+
+		await expect(starIcon1).toHaveClass(/bi-star-fill/);
+		await expect(starIcon2).toHaveClass(/bi-star(?!-fill)/);
+
+		// Unstar ds2 and reload page
+		await starButton1.click();
+
+		await expect(starIcon1).toHaveClass(/bi-star(?!-fill)/);
+		await expect(starIcon2).toHaveClass(/bi-star(?!-fill)/);
+
+		await page.reload();
+
+		// After reload:
+		// ds1 -> not starred
+		// ds2 -> not starred
+		dataset1 = datasetTable.getByRole('row').nth(1).getByRole('cell').nth(0);
+		dataset2 = datasetTable.getByRole('row').nth(2).getByRole('cell').nth(0);
+
+		await expect(dataset1).toHaveText(name1);
+		await expect(dataset2).toHaveText(name2);
+
+		starIcon1 = dataset1.getByRole('button', { name: 'star dataset' }).locator('i');
+		starIcon2 = dataset2.getByRole('button', { name: 'star dataset' }).locator('i');
+
+		await expect(starIcon1).toHaveClass(/bi-star(?!-fill)/);
+		await expect(starIcon2).toHaveClass(/bi-star(?!-fill)/);
 	});
 });